requirements.txt
Pillow
requests
tqdm
p_tqdm
